% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{get_agronomy}
\alias{get_agronomy}
\title{Get agronomy setup}
\usage{
get_agronomy(url_root)
}
\arguments{
\item{url_root}{Url root where the API is located.}
}
\value{
A List with 2 attributes (cultivars, soils).
}
\description{
function which gets a list of setups of cultivars and soils available into the aclimate platform.
}
\examples{
url_root = "https://pronosticosapi.aclimatecolombia.org/api/"
df = get_agronomy(url_root)
print(head(df))

}
