% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{get_forecast_crop}
\alias{get_forecast_crop}
\title{Get crop forecast}
\usage{
get_forecast_crop(url_root, stations)
}
\arguments{
\item{url_root}{Url root where the API is located.}

\item{stations}{Array of strings with the ids of the weather stations which want to search.}
}
\value{
A data.frame, with the crop forecast result for the weather stations.
}
\description{
function which gets the crop forecast for a set of weather stations available into the aclimate platform
}
\examples{
url_root = "https://pronosticosapi.aclimatecolombia.org/api/"
stations=c("58504f1a006cb93ed40eebe2","58504f1a006cb93ed40eebe3")
df = get_forecast_crop(url_root, stations)
print(head(df))

}
