% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{get_ws}
\alias{get_ws}
\title{Get weather stations}
\usage{
get_ws(url_root)
}
\arguments{
\item{url_root}{Url root where the API is located.}
}
\value{
A data.frame, with the list of all weather stations.
}
\description{
function which gets a list of all weather stations available into the aclimate platform.
}
\examples{
url_root = "https://pronosticosapi.aclimatecolombia.org/api/"
df_ws = get_ws(url_root)
print(head(df_ws))

}
